Added some functions
There are now:
createNewSlot() - not working YET -> adds a new slot to the "write atlas" panel

_show_edit_box() -- shows the edit box

readJson() -- literally read a json and add their information to the inventory system

writeJson() -- will literally write on the .json file. KEEP THIS IN MIND !! IT NEED TO FOLLOW THE ORIGINAL / MANDATORY FORMAT !!!!   PROBABLY I WILL ADD A "DEFAULT FORMAT BUTTON" BUT NOT SURE YET !!!!

und the _on_edit_properties() --- not implemented yet, but is triggered when clicked on "R_PROPERTIES_EDITOR" that should popup the "write atlas" vbox.